refactor(transform): move the resolved option structs beside their resolver
`features.rs` and `transform.test.ts` both went past the 350-line limit
for changed sources. The three small resolved-option structs belong next
to the resolver that builds them, and the edit-link tests are their own
subject.

Also refreshes the NAPI declarations snapshot for the new `srcDir`,
`provider`, and `urlPattern` fields.
